第一版测试基本完成，功能基本符合要求，没有严重BUG。
     可能post请求参数混乱，

似乎没有消息通知机制，也就是说点赞了也没红点提示....
反正还有三个月不急，早着呢。我要搞深度学习了

+++ src/main/groovy/cn/edu/ncepu/researchplatform/service/ScheduleService.java
@@ -79,16 +79,22 @@ public class ScheduleService {
     public void deleteExpireArticleFile() {
         LocalDateTime time = LocalDateTime.now().minusDays(retainDay);
         List<String> expiredArticle = articleMapper.findPathByDeleted(time);
-        expiredArticle.forEach(s -> Paths.get(pathPre, "ResearchPlatformFiles", "article",s).toFile().delete());
+        expiredArticle.forEach(s -> Paths.get(pathPre, "ResearchPlatformFiles", "article", s).toFile().delete());
         File[] tempFiles = Paths.get(pathPre, "ResearchPlatformFiles", "article").toFile().listFiles(pathname -> pathname.isFile() && pathname.getName().endsWith(".temp"));
+        if (tempFiles == null) {
+            return;
+        }
         for (File file : tempFiles) {
             file.delete();
         }
-}
+    }
 
     @Scheduled(cron = "@daily")
     public void deleteExpireMaterialFile() {
         File[] tempFiles = Paths.get(pathPre, "ResearchPlatformFiles", "material").toFile().listFiles(pathname -> pathname.isFile() && pathname.getName().endsWith(".temp"));
+        if (tempFiles == null) {
+            return;
+        }
         for (File file : tempFiles) {
             FileUtil.del(file);
         }
